implement graceful shutdown and wire up ActorContext to Actor.

+++ context.go
@@ -10,7 +10,6 @@ import (
 // Actor Context
 // TODO actor hierarchy and supervision (this would introduce actor system)
 type ActorContext struct {
-	System           *ActorSystem
 	Parent           *ActorContext
 	Self             *Actor
 	Children         set.Set
@@ -20,11 +19,9 @@ type ActorContext struct {
 	behaviorStack    []Receive
 	mailbox          chan Message
 	killChan         chan kill
-	terminateChan    chan terminate
 	attachMonChan    chan *Actor
 	detachMonChan    chan *Actor
 	addChildChan     chan *ActorContext
-	shutdownChan     chan shutdown
 	receiveTimeout   time.Duration
 }
 
@@ -55,31 +52,9 @@ func (context *ActorContext) Unbecome() {
 }
 
 // constructor
-func newTopLevelActorContext(system *ActorSystem, self *Actor, receive Receive) *ActorContext {
-	// TODO make parameters configurable
-	return &ActorContext{
-		System:           system,
-		Self:             self,
-		Children:         set.NewSet(),
-		originalBehavior: receive,
-		currentBehavior:  receive,
-		behaviorStack:    []Receive{receive},
-		mailbox:          make(chan Message, 100),
-		// buffer size for control message is 1 (cotrol method would block)
-		attachMonChan:  make(chan *Actor),
-		detachMonChan:  make(chan *Actor),
-		terminateChan:  make(chan terminate),
-		killChan:       make(chan kill),
-		addChildChan:   make(chan *ActorContext),
-		shutdownChan:   make(chan shutdown),
-		receiveTimeout: time.Duration(10) * time.Millisecond,
-	}
-}
-
 func newActorContext(parent *ActorContext, self *Actor, receive Receive) *ActorContext {
 	// TODO make parameters configurable
 	context := &ActorContext{
-		System:           parent.System,
 		Parent:           parent,
 		Self:             self,
 		Children:         set.NewSet(),
@@ -90,13 +65,13 @@ func newActorContext(parent *ActorContext, self *Actor, receive Receive) *ActorC
 		// buffer size for control message is 1 (cotrol method would block)
 		attachMonChan:  make(chan *Actor),
 		detachMonChan:  make(chan *Actor),
-		terminateChan:  make(chan terminate),
 		killChan:       make(chan kill),
 		addChildChan:   make(chan *ActorContext),
-		shutdownChan:   make(chan shutdown),
 		receiveTimeout: time.Duration(10) * time.Millisecond,
 	}
-	parent.addChildChan <- context
+	if parent != nil {
+		parent.addChildChan <- context
+	}
 	return context
 }
 
@@ -105,20 +80,22 @@ func (context *ActorContext) closeAllChan() {
 	go func() {
 		defer logPanic(context.Self)
 		close(context.mailbox)
-		close(context.terminateChan)
+		close(context.addChildChan)
+		close(context.attachMonChan)
+		close(context.detachMonChan)
 		close(context.killChan)
 	}()
 }
 
 func (context *ActorContext) start() chan bool {
 	startLatch := make(chan bool)
-	context.System.wg.Add(1)
+	context.Self.System.wg.Add(1)
 	go func() {
 		defer func() {
-			context.System.wg.Done()
+			context.Self.System.wg.Done()
 			context.closeAllChan()
 		}()
-		context.System.running.Add(context.Self)
+		context.Self.System.running.Add(context.Self)
 		<-startLatch
 		close(startLatch)
 		context.loop()
@@ -139,11 +116,7 @@ func (context *ActorContext) kill() {
 }
 
 func (context *ActorContext) terminate() {
-	context.terminateChan <- terminate{}
-}
-
-func (context *ActorContext) shutdown() {
-	context.shutdownChan <- shutdown{}
+	context.mailbox <- Message{PoisonPill{}}
 }
 
 // Actor's main loop which is executed in go routine
@@ -153,6 +126,7 @@ func (context *ActorContext) loop() {
 	// context.notifyMonitors(Message{STARTED})
 	for {
 		// TODO log
+		receiveTerminated := false
 		select {
 		case <-context.killChan:
 			context.notifyMonitors(Message{Down{
@@ -163,51 +137,50 @@ func (context *ActorContext) loop() {
 				context.Parent.kill()
 			}
 			return
-		case <-context.terminateChan:
-			context.notifyMonitors(Message{Down{
-				Cause: "terminated",
-				Actor: context.Self,
-			}})
-			if context.Parent != nil {
-				context.Parent.terminate()
-			}
-			return
 		case mon := <-context.attachMonChan:
 			if context.monitor == nil {
-				context.monitor = context.System.spawnMonitorForwarderFor(context.Self)
+				context.monitor = context.Self.System.spawnMonitorForwarderFor(context.Self)
 			}
 			context.monitor.Add(mon)
 		case mon := <-context.detachMonChan:
 			context.monitor.Remove(mon)
 		case child := <-context.addChildChan:
 			context.Children.Add(child)
-		case <-context.shutdownChan:
-			for child := range context.Children.Iter() {
-				if childContext, ok := child.(*ActorContext); ok {
-					childContext.shutdown()
-				}
-			}
-			context.notifyMonitors(Message{Down{
-				Cause: "shutdowned",
-				Actor: context.Self,
-			}})
-			return
 		default:
-			context.processOneMessage()
+			receiveTerminated = context.processOneMessage()
+		}
+		if receiveTerminated {
+			return
 		}
 		// force give back control to goroutine scheduler.
 		runtime.Gosched()
 	}
 }
 
-func (context *ActorContext) processOneMessage() {
+func (context *ActorContext) processOneMessage() bool{
 	select {
 	case msg := <-context.mailbox:
-		context.currentBehavior(msg, context)
+		if len(msg) == 1 {
+			if _, ok := msg[0].(PoisonPill); ok {
+				context.notifyMonitors(Message{Down{
+					Cause: "terminated",
+					Actor: context.Self,
+				}})
+				if context.Parent != nil {
+					context.Parent.terminate()
+				}
+				return true
+			} else {
+				context.currentBehavior(msg, context)
+			}
+		} else {
+			context.currentBehavior(msg, context)
+		}
 	case <-time.After(context.receiveTimeout):
 		// TODO log
 		// timeout
 	}
+	return false
 }
 
 func (context *ActorContext) notifyMonitors(msg Message) {
